package refactoring to move regex-related classes into the pattern.regex package
